requirements.txt
Django==1.7.4
Markdown==2.5.2
dj-database-url==0.3.0
dj-static==0.0.6
django-cors-headers==1.0.0
django-filter==0.9.2
django-toolbelt==0.0.1
djangorestframework==3.0.3
gunicorn==19.2.0
psycopg2==2.5.4
static3==0.5.1
wsgiref==0.1.2
